@@ -1,25 +1,24 @@
 name: flutter_instagram_stories
 homepage: https://github.com/awaik/flutter_instagram_stories
 description: Display stories (videos and images) like Whatsapp & Instagram.
-version: 0.1.5+7
+version: 1.0.0
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  rxdart: ^0.25.0
-  video_player: ^1.0.1
-  flutter_cache_manager: ^2.0.0
-  cached_network_image: ^2.5.0
-  json_annotation: ^3.1.0
-  firebase_core: ^0.5.3
-  cloud_firestore: ^0.14.3
-  dotted_border: ^1.0.7
-  sqflite: 1.3.2
+  rxdart: ^0.26.0
+  video_player: ^2.0.0
+  cached_network_image: ^3.0.0-nullsafety
+  firebase_core: ^1.0.0
+  cloud_firestore: ^1.0.0
+  dotted_border: ^2.0.0-nullsafety.0
+  json_annotation: ^4.0.0
+  collection: ^1.15.0-nullsafety.4
 
 dev_dependencies:
-  build_runner: ^1.10.3
-  json_serializable: ^3.5.0
-  test: ^1.15.4
+  build_runner: ^1.11.5
+  json_serializable: ^4.0.2
+  test:
