@@ -4,9 +4,9 @@ part 'story_data.g.dart';
 
 @JsonSerializable(explicitToJson: true)
 class StoryData {
-  String filetype;
-  Map<String, String> url;
-  Map<String, String> fileTitle;
+  String? filetype;
+  Map<String, String>? url;
+  Map<String, String>? fileTitle;
 
   StoryData({this.filetype, this.url});
 
