@@ -2,6 +2,7 @@ import 'dart:math';
 import 'dart:ui';
 import 'dart:async';
 
+import 'package:collection/collection.dart' show IterableExtension;
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'story_video.dart';
@@ -101,7 +102,7 @@ class StoryItem {
   static StoryItem pageImage(
     ImageProvider image, {
     BoxFit imageFit = BoxFit.fitWidth,
-    String caption,
+    String? caption,
     bool shown = false,
     Duration duration = const Duration(seconds: 3),
   }) {
@@ -158,18 +159,17 @@ class StoryItem {
   }
 
   static StoryItem pageGif(
-    String url, {
-    StoryController controller,
+    String? url, {
+    StoryController? controller,
     BoxFit imageFit = BoxFit.fitWidth,
-    String caption,
+    String? caption,
     bool shown = false,
     Duration duration = const Duration(seconds: 3),
-    Map<String, dynamic> requestHeaders,
-    TextStyle captionTextStyle,
-    EdgeInsets captionMargin,
-    EdgeInsets captionPadding,
+    Map<String, dynamic>? requestHeaders,
+    TextStyle? captionTextStyle,
+    EdgeInsets? captionMargin,
+    EdgeInsets? captionPadding,
   }) {
-    assert(imageFit != null, "[imageFit] should not be null");
     return StoryItem(
       Container(
         color: Colors.black,
@@ -212,17 +212,17 @@ class StoryItem {
   }
 
   static StoryItem pageVideo(
-    String url, {
-    StoryController controller,
+    String? url, {
+    StoryController? controller,
     //TODO: adjust duration to video length
     Duration duration = const Duration(seconds: 10),
     BoxFit imageFit = BoxFit.fitWidth,
     bool shown = false,
-    Map<String, dynamic> requestHeaders,
-    String caption,
-    TextStyle captionTextStyle,
-    EdgeInsets captionMargin,
-    EdgeInsets captionPadding,
+    Map<String, dynamic>? requestHeaders,
+    String? caption,
+    TextStyle? captionTextStyle,
+    EdgeInsets? captionMargin,
+    EdgeInsets? captionPadding,
   }) {
     assert(imageFit != null, "[imageFit] should not be null");
 
@@ -268,17 +268,17 @@ class StoryItem {
 /// gestures to pause, forward and go to previous page.
 class StoryView extends StatefulWidget {
   /// The pages to displayed.
-  final List<StoryItem> storyItems;
+  final List<StoryItem?> storyItems;
 
   /// Callback for when a full cycle of story is shown. This will be called
   /// each time the full story completes when [repeat] is set to `true`.
-  final VoidCallback onComplete;
+  final VoidCallback? onComplete;
 
   ///awaik
-  final VoidCallback goForward;
+  final VoidCallback? goForward;
 
   /// Callback for when a story is currently being shown.
-  final ValueChanged<StoryItem> onStoryShow;
+  final ValueChanged<StoryItem>? onStoryShow;
 
   /// Where the progress indicator should be placed.
   final ProgressPosition progressPosition;
@@ -291,7 +291,7 @@ class StoryView extends StatefulWidget {
   /// a [ListView] or [Column]) then set this to `true`.
   final bool inline;
 
-  final StoryController controller;
+  final StoryController? controller;
 
   StoryView(
     this.storyItems, {
@@ -318,14 +318,14 @@ class StoryView extends StatefulWidget {
 }
 
 class StoryViewState extends State<StoryView> with TickerProviderStateMixin {
-  AnimationController animationController;
-  Animation<double> currentAnimation;
-  Timer debouncer;
+  AnimationController? animationController;
+  Animation<double>? currentAnimation;
+  Timer? debouncer;
 
-  StreamSubscription<PlaybackState> playbackSubscription;
+  StreamSubscription<PlaybackState>? playbackSubscription;
 
-  StoryItem get lastShowing =>
-      widget.storyItems.firstWhere((it) => !it.shown, orElse: () => null);
+  StoryItem? get lastShowing =>
+      widget.storyItems.firstWhere((it) => !it!.shown, orElse: () => null);
 
   @override
   void initState() {
@@ -335,10 +335,10 @@ class StoryViewState extends State<StoryView> with TickerProviderStateMixin {
     // false
 
     final firstPage = widget.storyItems.firstWhere((it) {
-      return !it.shown;
+      return !it!.shown;
     }, orElse: () {
       widget.storyItems.forEach((it2) {
-        it2.shown = false;
+        it2!.shown = false;
       });
 
       return null;
@@ -347,7 +347,7 @@ class StoryViewState extends State<StoryView> with TickerProviderStateMixin {
     if (firstPage != null) {
       final lastShownPos = widget.storyItems.indexOf(firstPage);
       widget.storyItems.sublist(lastShownPos).forEach((it) {
-        it.shown = false;
+        it!.shown = false;
       });
     }
 
@@ -355,7 +355,7 @@ class StoryViewState extends State<StoryView> with TickerProviderStateMixin {
 
     if (widget.controller != null) {
       this.playbackSubscription =
-          widget.controller.playbackNotifier.listen((playbackStatus) {
+          widget.controller!.playbackNotifier.listen((playbackStatus) {
         if (playbackStatus == PlaybackState.play) {
           unpause();
         } else if (playbackStatus == PlaybackState.pause) {
@@ -384,17 +384,17 @@ class StoryViewState extends State<StoryView> with TickerProviderStateMixin {
     animationController?.dispose();
     // get the next playing page
     final storyItem = widget.storyItems.firstWhere((it) {
-      return !it.shown;
-    });
+      return !it!.shown;
+    })!;
 
     if (widget.onStoryShow != null) {
-      widget.onStoryShow(storyItem);
+      widget.onStoryShow!(storyItem);
     }
 
     animationController =
         AnimationController(duration: storyItem.duration, vsync: this);
 
-    animationController.addStatusListener((status) {
+    animationController!.addStatusListener((status) {
       if (status == AnimationStatus.completed) {
         storyItem.shown = true;
         if (widget.storyItems.last != storyItem) {
@@ -406,8 +406,9 @@ class StoryViewState extends State<StoryView> with TickerProviderStateMixin {
       }
     });
 
-    currentAnimation = Tween(begin: 0.0, end: 1.0).animate(animationController);
-    animationController.forward();
+    currentAnimation =
+        Tween(begin: 0.0, end: 1.0).animate(animationController!);
+    animationController!.forward();
   }
 
   void beginPlay() {
@@ -418,14 +419,14 @@ class StoryViewState extends State<StoryView> with TickerProviderStateMixin {
   void onComplete() {
     if (widget.onComplete != null) {
       widget.controller?.pause();
-      widget.onComplete();
+      widget.onComplete!();
     } else {
       print("Done");
     }
 
     if (widget.repeat) {
       widget.storyItems.forEach((it) {
-        it.shown = false;
+        it!.shown = false;
       });
 
       beginPlay();
@@ -435,18 +436,18 @@ class StoryViewState extends State<StoryView> with TickerProviderStateMixin {
   void goBack() {
     widget.controller?.play();
 
-    animationController.stop();
+    animationController!.stop();
 
     if (this.lastShowing == null) {
-      widget.storyItems.last.shown = false;
+      widget.storyItems.last!.shown = false;
     }
 
     if (this.lastShowing == widget.storyItems.first) {
       beginPlay();
     } else {
-      this.lastShowing.shown = false;
+      this.lastShowing!.shown = false;
       int lastPos = widget.storyItems.indexOf(this.lastShowing);
-      final previous = widget.storyItems[lastPos - 1];
+      final previous = widget.storyItems[lastPos - 1]!;
 
       previous.shown = false;
 
@@ -456,7 +457,7 @@ class StoryViewState extends State<StoryView> with TickerProviderStateMixin {
 
   void goForward() {
     if (this.lastShowing != widget.storyItems.last) {
-      animationController.stop();
+      animationController!.stop();
 
       // get last showing
       final _last = this.lastShowing;
@@ -469,7 +470,7 @@ class StoryViewState extends State<StoryView> with TickerProviderStateMixin {
       }
     } else {
       // this is the last page, progress animation should skip to end
-      animationController.animateTo(1.0, duration: Duration(milliseconds: 10));
+      animationController!.animateTo(1.0, duration: Duration(milliseconds: 10));
     }
   }
 
@@ -483,7 +484,7 @@ class StoryViewState extends State<StoryView> with TickerProviderStateMixin {
 
   void controlPause() {
     if (widget.controller != null) {
-      widget.controller.pause();
+      widget.controller!.pause();
     } else {
       pause();
     }
@@ -491,14 +492,14 @@ class StoryViewState extends State<StoryView> with TickerProviderStateMixin {
 
   void controlUnpause() {
     if (widget.controller != null) {
-      widget.controller.play();
+      widget.controller!.play();
     } else {
       unpause();
     }
   }
 
   Widget get currentView => widget.storyItems
-      .firstWhere((it) => !it.shown, orElse: () => widget.storyItems.last)
+      .firstWhere((it) => !it!.shown, orElse: () => widget.storyItems.last)!
       .view;
 
   @override
@@ -522,7 +523,7 @@ class StoryViewState extends State<StoryView> with TickerProviderStateMixin {
                 ),
                 child: PageBar(
                   widget.storyItems
-                      .map((it) => PageData(it.duration, it.shown))
+                      .map((it) => PageData(it!.duration, it.shown))
                       .toList(),
                   this.currentAnimation,
                   key: UniqueKey(),
@@ -554,13 +555,13 @@ class StoryViewState extends State<StoryView> with TickerProviderStateMixin {
                     ..onSecondaryTapUp = (details) {
                       if (debouncer?.isActive == true) {
                         print('+++onTapUp11');
-                        debouncer.cancel();
+                        debouncer!.cancel();
                         debouncer = null;
 //                        goForward();
                         controlUnpause();
                       } else {
                         print('+++onTapUp21');
-                        debouncer.cancel();
+                        debouncer!.cancel();
                         debouncer = null;
 
                         controlUnpause();
@@ -569,13 +570,13 @@ class StoryViewState extends State<StoryView> with TickerProviderStateMixin {
                     ..onTapCancel = () {
                       if (debouncer?.isActive == true) {
                         print('+++onTapUp12');
-                        debouncer.cancel();
+                        debouncer!.cancel();
                         debouncer = null;
 //                        goForward();
                         controlUnpause();
                       } else {
                         print('+++onTapUp22');
-                        debouncer.cancel();
+                        debouncer!.cancel();
                         debouncer = null;
 
                         controlUnpause();
@@ -584,13 +585,13 @@ class StoryViewState extends State<StoryView> with TickerProviderStateMixin {
                     ..onTapUp = (details) {
                       if (debouncer?.isActive == true) {
                         print('+++onTapUp13');
-                        debouncer.cancel();
+                        debouncer!.cancel();
                         debouncer = null;
 //                        goForward();
                         controlUnpause();
                       } else {
                         print('+++onTapUp23');
-                        debouncer.cancel();
+                        debouncer!.cancel();
                         debouncer = null;
 
                         controlUnpause();
@@ -631,14 +632,14 @@ class PageData {
 /// [pages] provided.
 class PageBar extends StatefulWidget {
   final List<PageData> pages;
-  final Animation<double> animation;
+  final Animation<double>? animation;
   final IndicatorHeight indicatorHeight;
 
   PageBar(
     this.pages,
     this.animation, {
     this.indicatorHeight = IndicatorHeight.large,
-    Key key,
+    Key? key,
   }) : super(key: key);
 
   @override
@@ -661,7 +662,7 @@ class PageBarState extends State<PageBar> {
             ? 2
             : 4;
 
-    widget.animation.addListener(() {
+    widget.animation!.addListener(() {
       setState(() {});
     });
   }
@@ -674,8 +675,7 @@ class PageBarState extends State<PageBar> {
   }
 
   bool isPlaying(PageData page) {
-    return widget.pages.firstWhere((it) => !it.shown, orElse: () => null) ==
-        page;
+    return widget.pages.firstWhereOrNull((it) => !it.shown) == page;
   }
 
   @override
@@ -688,7 +688,7 @@ class PageBarState extends State<PageBar> {
                 right: widget.pages.last == it ? 0 : this.spacing),
             child: StoryProgressIndicator(
               isPlaying(it)
-                  ? widget.animation.value
+                  ? widget.animation!.value
                   : it.shown
                       ? 1
                       : 0,
@@ -757,9 +757,9 @@ class IndicatorOval extends CustomPainter {
 
 /// Concept source: https://stackoverflow.com/a/9733420
 class ContrastHelper {
-  static double luminance(int r, int g, int b) {
+  static double luminance(int? r, int? g, int? b) {
     final a = [r, g, b].map((it) {
-      double value = it.toDouble() / 255.0;
+      double value = it!.toDouble() / 255.0;
       return value <= 0.03928
           ? value / 12.92
           : pow((value + 0.055) / 1.055, 2.4);
