@@ -8,14 +8,13 @@ part of 'stories.dart';
 
 Stories _$StoriesFromJson(Map<String, dynamic> json) {
   return Stories(
-    storyId: json['storyId'] as String,
+    storyId: json['storyId'] as String?,
     date: json['date'] == null ? null : DateTime.parse(json['date'] as String),
-    file: (json['file'] as List)
-        ?.map((e) =>
-            e == null ? null : StoryData.fromJson(e as Map<String, dynamic>))
-        ?.toList(),
-    previewImage: json['previewImage'] as String,
-    previewTitle: (json['previewTitle'] as Map<String, dynamic>)?.map(
+    file: (json['file'] as List<dynamic>?)
+        ?.map((e) => StoryData.fromJson(e as Map<String, dynamic>))
+        .toList(),
+    previewImage: json['previewImage'] as String?,
+    previewTitle: (json['previewTitle'] as Map<String, dynamic>?)?.map(
       (k, e) => MapEntry(k, e as String),
     ),
   );
@@ -24,7 +23,7 @@ Stories _$StoriesFromJson(Map<String, dynamic> json) {
 Map<String, dynamic> _$StoriesToJson(Stories instance) => <String, dynamic>{
       'storyId': instance.storyId,
       'date': instance.date?.toIso8601String(),
-      'file': instance.file?.map((e) => e?.toJson())?.toList(),
+      'file': instance.file?.map((e) => e.toJson()).toList(),
       'previewImage': instance.previewImage,
       'previewTitle': instance.previewTitle,
     };
