@@ -8,11 +8,11 @@ part of 'story_data.dart';
 
 StoryData _$StoryDataFromJson(Map<String, dynamic> json) {
   return StoryData(
-    filetype: json['filetype'] as String,
-    url: (json['url'] as Map<String, dynamic>)?.map(
+    filetype: json['filetype'] as String?,
+    url: (json['url'] as Map<String, dynamic>?)?.map(
       (k, e) => MapEntry(k, e as String),
     ),
-  )..fileTitle = (json['fileTitle'] as Map<String, dynamic>)?.map(
+  )..fileTitle = (json['fileTitle'] as Map<String, dynamic>?)?.map(
       (k, e) => MapEntry(k, e as String),
     );
 }
